detection_program/requirements.txt
pytorch
torchvision
visdom
cudatoolkit=10.1
opencv
sklearn
ujson
matplotlib